mgr/dashboard: improve edit site name action in rbd-mirroring
Fixes: https://tracker.ceph.com/issues/55896
