--- src/pybind/mgr/dashboard/frontend/src/app/ceph/block/mirroring/edit-site-name-modal/edit-site-name-modal.component.html
@@ -1,39 +1,0 @@
-<cd-modal [modalRef]="activeModal">
-  <ng-container i18n
-                class="modal-title">Edit site name</ng-container>
-
-  <ng-container class="modal-content">
-    <form name="editSiteNameForm"
-          class="form"
-          #formDir="ngForm"
-          [formGroup]="editSiteNameForm"
-          novalidate>
-      <div class="modal-body">
-        <p>
-          <ng-container i18n>Edit the site name and click&nbsp;
-          <kbd>Update</kbd>.</ng-container>
-        </p>
-
-        <div class="form-group">
-          <label class="col-form-label required"
-                 for="siteName"
-                 i18n>Site Name</label>
-          <input class="form-control"
-                 type="text"
-                 placeholder="Name..."
-                 i18n-placeholder
-                 id="siteName"
-                 name="siteName"
-                 formControlName="siteName"
-                 autofocus>
-        </div>
-      </div>
-
-      <div class="modal-footer">
-        <cd-form-button-panel (submitActionEvent)="update()"
-                              [form]="editSiteNameForm"
-                              [submitText]="actionLabels.UPDATE"></cd-form-button-panel>
-      </div>
-    </form>
-  </ng-container>
-</cd-modal>
