+++ src/pybind/mgr/dashboard/frontend/src/app/ceph/block/mirroring/overview/overview.component.spec.ts
@@ -1,11 +1,14 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ReactiveFormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
 
 import { NgbNavModule, NgbProgressbarModule } from '@ng-bootstrap/ng-bootstrap';
 import { ToastrModule } from 'ngx-toastr';
+import { of } from 'rxjs';
 
+import { RbdMirroringService } from '~/app/shared/api/rbd-mirroring.service';
 import { SharedModule } from '~/app/shared/shared.module';
 import { configureTestBed } from '~/testing/unit-test-helper';
 import { DaemonListComponent } from '../daemon-list/daemon-list.component';
@@ -17,6 +20,7 @@ import { OverviewComponent } from './overview.component';
 describe('OverviewComponent', () => {
   let component: OverviewComponent;
   let fixture: ComponentFixture<OverviewComponent>;
+  let rbdMirroringService: RbdMirroringService;
 
   configureTestBed({
     declarations: [
@@ -33,6 +37,7 @@ describe('OverviewComponent', () => {
       NgbProgressbarModule,
       HttpClientTestingModule,
       RouterTestingModule,
+      ReactiveFormsModule,
       ToastrModule.forRoot()
     ]
   });
@@ -40,10 +45,35 @@ describe('OverviewComponent', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(OverviewComponent);
     component = fixture.componentInstance;
+    rbdMirroringService = TestBed.inject(RbdMirroringService);
+    component.siteName = 'site-A';
     fixture.detectChanges();
   });
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  describe('edit site name', () => {
+    beforeEach(() => {
+      spyOn(rbdMirroringService, 'getSiteName').and.callFake(() => of({ site_name: 'site-A' }));
+      spyOn(rbdMirroringService, 'refresh').and.stub();
+      fixture.detectChanges();
+    });
+
+    afterEach(() => {
+      expect(rbdMirroringService.refresh).toHaveBeenCalledTimes(1);
+    });
+
+    it('should call setSiteName', () => {
+      component.editing = true;
+      spyOn(rbdMirroringService, 'setSiteName').and.callFake(() => of({ site_name: 'new-site-A' }));
+
+      component.rbdmirroringForm.patchValue({
+        siteName: 'new-site-A'
+      });
+      component.updateSiteName();
+      expect(rbdMirroringService.setSiteName).toHaveBeenCalledWith('new-site-A');
+    });
   });
 });
